Rename ccli to ctl (and libccli to ctltools)
