+++ ctltools/clean.sh
