+++ ctltools/smtlib-compare.sh
