+++ ctltools/smtlib-dump.sh
