+++ ctltools/smtlib-sort.sh
